backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
tldextract==5.3.0
python-whois==0.9.6
